fix: siem migration privs
