+++ x-pack/solutions/security/plugins/security_solution/public/siem_migrations/rules/service/rule_migrations_service.ts
@@ -6,6 +6,8 @@
  */
 
 import type { CoreStart } from '@kbn/core/public';
+import { licenseService } from '../../../common/hooks/use_license';
+import { ExperimentalFeaturesService } from '../../../common/experimental_features_service';
 import type { TelemetryServiceStart } from '../../../common/lib/telemetry';
 import type {
   CreateRuleMigrationRulesRequestBody,
@@ -20,14 +22,17 @@ import * as api from '../api';
 import type { RuleMigrationSettings, RuleMigrationStats } from '../types';
 import * as i18n from './translations';
 import { SiemRulesMigrationsTelemetry } from './telemetry';
+import type { CapabilitiesLevel, MissingCapability } from '../../common/service';
 import {
   SiemMigrationsServiceBase,
+  getMissingCapabilitiesChecker,
   getMissingCapabilitiesToast,
   getNoConnectorToast,
 } from '../../common/service';
 import type { GetMigrationStatsParams, GetMigrationsStatsAllParams } from '../../common/types';
 import { getSuccessToast } from './notification/success_notification';
 import { START_STOP_POLLING_SLEEP_SECONDS } from '../../common/constants';
+import { requiredRuleMigrationCapabilities } from './capabilities';
 
 const CREATE_MIGRATION_BODY_BATCH_SIZE = 50;
 
@@ -41,6 +46,22 @@ export class SiemRulesMigrationsService extends SiemMigrationsServiceBase<RuleMi
   ) {
     super(core, plugins);
     this.telemetry = new SiemRulesMigrationsTelemetry(telemetryService);
+  }
+
+  /** Returns any missing capabilities for the user to use this feature */
+  public getMissingCapabilities(level?: CapabilitiesLevel): MissingCapability[] {
+    const getMissingCapabilities = getMissingCapabilitiesChecker(requiredRuleMigrationCapabilities);
+    return getMissingCapabilities(this.core.application.capabilities, level);
+  }
+
+  /** Checks if the service is available based on the `license`, `capabilities` and `experimentalFeatures` */
+  public isAvailable() {
+    const { siemMigrationsDisabled } = ExperimentalFeaturesService.get();
+    return (
+      !siemMigrationsDisabled &&
+      licenseService.isEnterprise() &&
+      !this.hasMissingCapabilities('minimum')
+    );
   }
 
   /** Accessor for the rule migrations API client */
